Update jquery.colorbox.js
Added ability to assign a title attribute to the generated iFrame element.  Use by setting the data-cbox-ititle="My iFrame title" attribute in the calling element.  Renders as: 

<iframe ... title="My iFrame title"></iframe>

I chose data-cbox-ititle instead of data-cbox-title just in case "title" might be used somewhere down the line.

This was necessary to make the iFrame pop-ups 508 compliant.